Add tests, major cleanup

+++ src/main/java/com/alexeygold2077/taskdeck/model/dto/ProjectDTO.java
@@ -12,7 +12,7 @@ public class ProjectDTO {
     private Long id;
     private String name;
     private String description;
-    private Integer createdAt;
+    private Long createdAt;
 
     private Long creatorId;
 }
